Corrected GEAR LEDs sequence in Manual LED mode.
In a real aircraft, the lights are NOT illuminated when the gear is UP.
However in the 'Manual Mode' they stayed RED.

RED should only show when gear is 'transitioning', so now, when the
level is changed, the LEDS change instantly to RED, and the either go
GREEN or DARK depending on the levers position.

This has been tested and checked for correct operation even when the
level is changed mid-transition.

This solution means that people using this brilliant app, but with games
other than DCS, can now accurately simulate gear transitions e.g. Star
Citizen and Elite Dangerous.

Thanks go to jdahlblom for his fantastic efforts, very clean, readable
and elegant code.